feat: 페이지 라우팅을 변경하라
- /about 페이지 -> / 페이지로 변경
- / 페이지 -> /maps 페이지로 변경
- 테스트 수정
