--- app/about/layout.tsx
@@ -1,17 +1,0 @@
-import { ReactNode } from 'react';
-
-import SearchCountryHeader from 'components/about/searchCountryHeader';
-
-type Props = {
-  children: ReactNode;
-};
-
-function Layout({ children }: Props) {
-  return (
-    <SearchCountryHeader>
-      {children}
-    </SearchCountryHeader>
-  );
-}
-
-export default Layout;
