+++ app/maps/page.tsx
@@ -1,0 +1,13 @@
+import MainMap from 'components/map/mainMap';
+
+export const metadata = {
+  title: 'offbeat-map',
+};
+
+function Page() {
+  return (
+    <MainMap />
+  );
+}
+
+export default Page;
