+++ components/about/countryItem/index.test.tsx
@@ -1,9 +1,24 @@
-import { render } from '@testing-library/react';
+import { useRouter } from 'next/navigation';
+
+import { fireEvent, render, screen } from '@testing-library/react';
 
 import CountryItem from '.';
 
+jest.mock('next/navigation', () => ({
+  useRouter: jest.fn(),
+}));
+
 describe('CountryItem', () => {
   const countryName = '대한민국';
+  const mockPush = jest.fn();
+
+  beforeEach(() => {
+    jest.clearAllMocks();
+
+    (useRouter as jest.Mock).mockImplementation(() => ({
+      push: mockPush,
+    }));
+  });
 
   const renderCountryItem = () => render((
     <CountryItem emoji="🇰🇷" koreanName={countryName} />
@@ -13,5 +28,15 @@ describe('CountryItem', () => {
     const { container } = renderCountryItem();
 
     expect(container).toHaveTextContent(countryName);
+  });
+
+  describe('나라 아이템을 선택한다', () => {
+    it('router push 이벤트가 발생해야만 한다', () => {
+      renderCountryItem();
+
+      fireEvent.click(screen.getByText(countryName));
+
+      expect(mockPush).toHaveBeenCalledWith('/maps');
+    });
   });
 });
